refactor(internal-account): extract InitiateLien helpers to separate file
Move InitiateLien helper functions from lien_service.go (821 lines,
over 800-line limit) to lien_initiate_helpers.go. Fixes Service
Conventions CI check.
